refactor: redux관련 파일 이름 변경 및 확장자 변경
