+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
 import { Provider } from "react-redux";
-import { store } from "./store/AppStore.jsx";
+import { store } from "./store/index.js";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
